requirements.txt
discord.py>=2.3.2
aiosqlite>=0.19.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
